==> src/2_process_vitals_labs_R.ipynb <==
library(tidyverse)
library(data.table)

# Load the cohort and the interval times -- interval times derived in 'create_action_space.ipynb'
data_path <- '../data/'
cohort_df <- read_csv(paste0(data_path, 'cohort.csv'))
interval_times_df <- read_csv(paste0(data_path, 'interval_times_df.csv'))

# Load the vitals
vitals_df <- read_csv(paste0(data_path, 'vitals_cohort.csv'))

head(vitals_df)

# Cast to DT
vitals_dt <- as.data.table(vitals_df, key = 'icustay_id')
interval_times_dt <- as.data.table(interval_times_df, key = 'icustay_id')

# 1:m merge everything together by icustay_id; long format
merged_vitals <- interval_times_dt %>% 
                    merge(vitals_dt, allow.cartesian = TRUE)

# only keep rows where vital time aligns with interval time; retain mean vital value for that timeframe
mean_vitals <- merged_vitals[charttime > interval_start_time & charttime <= interval_end_time,
                                   .(mean_vital = mean(valuenum)), by = .(subject_id, 
                                                                          hadm_id, 
                                                                          icustay_id, 
                                                                          interval_start_time, 
                                                                          interval_end_time, 
                                                                          vital_id)]

head(mean_vitals)

labs_df <- read_csv(paste0(data_path, 'labs_cohort.csv'))

labs_dt <- as.data.table(labs_df, key = 'icustay_id')

merged_labs <- interval_times_dt %>% 
                    merge(labs_dt, allow.cartesian = TRUE)

mean_labs <- merged_labs[charttime > interval_start_time & charttime <= interval_end_time,
                                   .(mean_lab = mean(valuenum)), by = .(subject_id, 
                                                                          hadm_id, 
                                                                          icustay_id, 
                                                                          interval_start_time, 
                                                                          interval_end_time, 
                                                                          lab_id)]

mean_labs <- mean_labs %>% rename(meas_id = lab_id, mean_value = mean_lab)
mean_vitals <- mean_vitals %>% rename(meas_id = vital_id, mean_value = mean_vital)

mean_labs_vitals <- mean_labs %>% full_join(mean_vitals)

vitals_labs_spread <- mean_labs_vitals %>% 
                        spread(meas_id, mean_value) %>% 
                        right_join(interval_times_dt) 

# Exclude times in which no measurements were made
vitals_labs_spread_filt <- vitals_labs_spread %>%
                                gather(lab_id, meas_value, ALBUMIN:WBC) %>%
                                group_by(subject_id, hadm_id, icustay_id, interval_start_time, interval_end_time) %>%
                                summarise(exclude = all(is.na(meas_value))) %>%
                                full_join(vitals_labs_spread) %>%
                                filter(!exclude)

head(vitals_labs_spread_filt)

write_csv(vitals_labs_spread_filt, paste0(data_path, 'vitals_labs_spread_filt.csv'))
