==> src/3_split_data_and_impute_R.ipynb <==
library(tidyverse)

data_path <- '../data/'
vitals_labs <- read_csv(paste0(data_path, 'vitals_labs_spread_filt.csv'))
demographics <- read_csv(paste0(data_path, 'demographics_cohort.csv'))
action_df <- read_csv(paste0(data_path, 'action_df.csv'))

names(vitals_labs)

# Perform LVCF
vitals_labs_lvcf <- vitals_labs %>% group_by(subject_id, hadm_id, icustay_id) %>%
                        arrange(subject_id, hadm_id, icustay_id, interval_start_time) %>%
                        fill(ALBUMIN:WBC)

data_all <- vitals_labs_lvcf %>% full_join(demographics) %>% ungroup()

# Generate the train/test split
set.seed(10)
ids <- unique(data_all$icustay_id)

train_prop <- 0.8
num_train <- floor(train_prop * length(ids))

train_ids <- sample(ids, num_train)
test_ids <- setdiff(ids, train_ids)

# Proportion validation - of the test data
val_prop <- 0.5 # 0.5 of 0.2 = 0.1 total
num_val <- floor(val_prop * length(test_ids))
val_ids <- sample(test_ids, num_val)
test_ids <- setdiff(test_ids, val_ids)

train_id_df <- data.frame(icustay_id = train_ids)
val_id_df <- data.frame(icustay_id = val_ids)
test_id_df <- data.frame(icustay_id = test_ids)

train_data <- data_all %>% inner_join(train_id_df) %>% ungroup()
val_data <- data_all %>% inner_join(val_id_df) %>% ungroup()
test_data <- data_all %>% inner_join(test_id_df) %>% ungroup()

dim(train_data)
dim(val_data)
dim(test_data)

library(caret)

feature_names <- setdiff(c(names(vitals_labs_lvcf), names(demographics)),
                         c('subject_id', 
                           'hadm_id', 
                           'icustay_id', 
                           'interval_start_time',
                           'interval_end_time',
                           'exclude'))
train_features <- train_data[, feature_names]
val_features <- val_data[, feature_names]
test_features <- test_data[, feature_names]

# Preprocess the data

preprocessor <- preProcess(train_features, method = c('center', 'scale', 'medianImpute'))
train_proc <- predict(preprocessor, train_features)
val_proc <- predict(preprocessor, val_features)
test_proc <- predict(preprocessor, test_features)

id_vars <- c('subject_id', 
               'hadm_id', 
               'icustay_id', 
               'interval_start_time',
               'interval_end_time')

# train_out <- cbind(train_proc, train_data[, !(names(train_data) %in% names(train_proc))])

# test_out <- cbind(test_proc, test_data[, !(names(test_data) %in% names(test_proc))])

train_out <- cbind(train_data[, !(names(train_data) %in% names(train_proc))], train_proc)
val_out <- cbind(val_data[, !(names(val_data) %in% names(val_proc))], val_proc)
test_out <- cbind(test_data[, !(names(test_data) %in% names(test_proc))], test_proc)


# train_proc[, id_features] <- train_data[, id_vars]
# test_proc[ ,id_features] <- test_data[, id_vars]
# head(train_proc, id_features)

head(train_out)

dim(train_out)
dim(val_out)
dim(test_out)

write_csv(train_out, paste0(data_path, 'train_data.csv'))
write_csv(val_out, paste0(data_path, 'val_data.csv'))
write_csv(test_out, paste0(data_path, 'test_data.csv'))
