==> src/4_prepare_df_for_tuplizing_R.ipynb <==
library(tidyverse)

data_path <- '../data/'
cohort_df <- read_csv(paste0(data_path, 'cohort.csv'))
action_df <- read_csv(paste0(data_path, 'action_df.csv'))
train_df <- read_csv(paste0(data_path, 'train_data.csv'))
val_df <- read_csv(paste0(data_path, 'val_data.csv'))
test_df <- read_csv(paste0(data_path, 'test_data.csv'))

outcome_df <- cohort_df %>% group_by(subject_id, 
                                     hadm_id, 
                                     icustay_id)

train_df_action <- train_df %>% inner_join(select(action_df, subject_id, 
                                                   hadm_id, 
                                                   icustay_id, 
                                                   interval_start_time, 
                                                   interval_end_time,
                                                   discrete_action)) %>%
                                inner_join(cohort_df)

val_df_action <- val_df %>% inner_join(select(action_df, subject_id, 
                                                   hadm_id, 
                                                   icustay_id, 
                                                   interval_start_time, 
                                                   interval_end_time,
                                                   discrete_action)) %>%
                                inner_join(cohort_df)

test_df_action <- test_df %>% inner_join(select(action_df, subject_id, 
                                                   hadm_id, 
                                                   icustay_id, 
                                                   interval_start_time, 
                                                   interval_end_time,
                                                   discrete_action)) %>%
                                inner_join(cohort_df)

temp_train <- train_df_action %>% group_by(subject_id, 
                             hadm_id, 
                             icustay_id,
                             hospital_expire_flag) %>%
                    summarise(interval_start_time = max(interval_start_time)) %>%
                    mutate(reward = ifelse(hospital_expire_flag, -15, 15))

train_df_action_reward <- train_df_action %>% 
                                left_join(temp_train) %>%
                                replace_na(list(reward = 0))

temp_val <- val_df_action %>% group_by(subject_id, 
                             hadm_id, 
                             icustay_id,
                             hospital_expire_flag) %>%
                    summarise(interval_start_time = max(interval_start_time)) %>%
                    mutate(reward = ifelse(hospital_expire_flag, -15, 15))

val_df_action_reward <- val_df_action %>% 
                                left_join(temp_val) %>%
                                replace_na(list(reward = 0))

temp_test <- test_df_action %>% group_by(subject_id, 
                             hadm_id, 
                             icustay_id,
                             hospital_expire_flag) %>%
                    summarise(interval_start_time = max(interval_start_time)) %>%
                    mutate(reward = ifelse(hospital_expire_flag, -15, 15))

test_df_action_reward <- test_df_action %>% 
                                left_join(temp_test) %>%
                                replace_na(list(reward = 0))

temp2_train <- train_df_action_reward %>% 
                                mutate(row_id = 1:nrow(train_df_action_reward) - 1) %>% 
                                arrange(subject_id, hadm_id, icustay_id, interval_start_time, row_id) %>%
                                group_by(subject_id, hadm_id, icustay_id) %>%
                                mutate(row_id_next = ifelse(row_id == max(row_id), row_id, row_id + 1))

temp2_val <- val_df_action_reward %>% 
                                mutate(row_id = 1:nrow(val_df_action_reward) - 1) %>% 
                                arrange(subject_id, hadm_id, icustay_id, interval_start_time, row_id) %>%
                                group_by(subject_id, hadm_id, icustay_id) %>%
                                mutate(row_id_next = ifelse(row_id == max(row_id), row_id, row_id + 1))

temp2_test <- test_df_action_reward %>% 
                                mutate(row_id = 1:nrow(test_df_action_reward) - 1) %>% 
                                arrange(subject_id, hadm_id, icustay_id, interval_start_time, row_id) %>%
                                group_by(subject_id, hadm_id, icustay_id) %>%
                                mutate(row_id_next = ifelse(row_id == max(row_id), row_id, row_id + 1))


final_df_train <- temp2_train %>% select(
                             -exclude, 
                             -intime, 
                             -outtime, 
                             -window_start,
                             -window_end,
                             -suspected_infection_time_poe
                             )

final_df_val <- temp2_val %>% select(
                             -exclude, 
                             -intime, 
                             -outtime, 
                             -window_start,
                             -window_end,
                             -suspected_infection_time_poe
                             )

final_df_test <- temp2_test %>% select(
                             -exclude, 
                             -intime, 
                             -outtime, 
                             -window_start,
                             -window_end,
                             -suspected_infection_time_poe
                             )

head(final_df_train)

names(final_df_train)
names(final_df_test)

write_csv(final_df_train, paste0(data_path, 'train_state_action_reward_df.csv'))
write_csv(final_df_val, paste0(data_path, 'val_state_action_reward_df.csv'))
write_csv(final_df_test, paste0(data_path, 'test_state_action_reward_df.csv'))
